Update timing demo benchmarking scripts to launch multiple SLURM jobs using a multiprocessing pool so that they can run in parallel if there are enough GPUs available.

+++ examples/timing/run_gpu_slow.py
@@ -6,6 +6,7 @@ Run timing test (GPU) scaled over number of ports.
 
 import csv
 import glob
+import multiprocessing as mp
 import os
 import re
 import subprocess
@@ -13,13 +14,29 @@ import sys
 
 import numpy as np
 
+from neurokernel.tools.misc import get_pids_open
+
+try:
+    from subprocess import DEVNULL
+except ImportError:
+    import os
+    DEVNULL = open(os.devnull, 'wb')
+
 out_file = sys.argv[1]
 script_name = 'timing_demo_gpu_slow.py'
 trials = 3
 lpus = 2
 
-f = open(out_file, 'w', 0)
-w = csv.writer(f)
+def check_and_print_output(*args):
+    try:
+        out = subprocess.check_output(*args, env=os.environ, stderr=DEVNULL)
+    except Exception as e:
+        out = e.output
+    print out,
+    return out
+
+pool = mp.Pool(5)
+results = []
 for spikes in np.linspace(50, 15000, 25, dtype=int):
     for i in xrange(trials):
         # CUDA < 7.0 doesn't properly clean up IPC-related files; since
@@ -27,15 +44,20 @@ for spikes in np.linspace(50, 15000, 25, dtype=int):
         # each job:
         ipc_files = glob.glob('/dev/shm/cuda.shm*')
         for ipc_file in ipc_files:
-            os.remove(ipc_file)
-        out = subprocess.check_output(['srun', '-n', '1', '-c', str(lpus+2),
-                                       '-p', 'huxley',
-                                       '--gres=gpu:%s' % lpus,
-                                       'python', script_name,
-                                       '-u', str(lpus), '-s', str(spikes),
-                                       '-g', '0', '-m', '50'])
-        average_step_sync_time, runtime_all, runtime_main, \
-            runtime_loop = out.strip('()\n\"').split(', ')
-        w.writerow([lpus, spikes, average_step_sync_time,
-                    runtime_all, runtime_main, runtime_loop])
+
+            # Only remove files that are not being held open by any processes:
+            if not get_pids_open(ipc_file):
+                os.remove(ipc_file)
+        r = pool.apply_async(check_and_print_output, 
+                             [['srun', '-n', '1', '-c', str(lpus+2),
+                               '-p', 'huxley',
+                               '--gres=gpu:%s' % lpus,
+                               'python', script_name,
+                               '-u', str(lpus), '-s', str(spikes),
+                               '-g', '0', '-m', '50']])
+        results.append(r)
+f = open(out_file, 'w', 0)
+w = csv.writer(f)
+for r in results:
+    w.writerow(r.get().strip('[]\n\"').split(', '))
 f.close()
